Don't use config file_list when file args supplied
Only use the file_list from the JSON configuration if no filename
arguments are supplied.

This allows local_rules and rule settings from a common configuration
file to be used when analysing only a single file.

This is particularly useful when VSG is being used as a checker by an
editor plugin such as Syntastic.