Add type and constant to the block unit test
Add an array type and constant to the block unit test input file to
increase the range of scenarios covered.

Change the entity name from "ENTITY" to "BLOCK_EXAMPLE" so that it does
not conflict with the keyword.

Add an entity declaration so that the file is a valid standalone VHDL
design.