Simplify block test
Use list comprehensions rather than iterating and appending.

Don't worry about limiting the number of lines considered in
test_block_indent_assignment(). The other tests don't do this and the
comparison will work as expected even if the lists are different sizes.

Remove unnecessary comments. It is obvious that the result of a list
comprehension will be a list and that assertEqual will compare its two
arguments.