Updates for issue #486 and #461:
When requesting version using master, it would report the previous released version.
It would also report the SHA of the GIT repository it was currently executing in.

Both of these issues make it unclear which version of VSG you are using.

This update provides

  1)  Defaulting version to 'master'
  2)  Defaulting SHA to NA

The goal will be to hardcode the version and SHA when releasing to PYPI.