Updates for issue #486:
Fixing version reporting with Git SHA.

  1)  Updating functions to version.py
      a)  get_git_sha
      b)  get_git_branch
  2)  Updated setup.py
      a)  Uses/generates file vsg/version_info.py depending on if it exists

NOTES:

version.py attempts to pull the latest version information from vsg/version_info.py.
If it exists it will use the version and sha in the file.
If it does not exist, version.py will pull the git sha from the local git repo.

When installing using setup.py, if the vsg/version_info.py exists then the version and sha in the file will be used.
If the file does not exist, then setup.py will pull the branch and sha from the git repo and build a unique version.